refactor header w/ dark mode
